+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - `ForeignStagingTableOperator` will now error if the underlying file cannot be read or the command returns an error. [#5763](https://github.com/Flowminder/FlowKit/issues/5763)
+- Flowmachine now requires SQLAlchemy >= 2.0.0 [#6066](https://github.com/Flowminder/FlowKit/issues/6066)
 
 ### Fixed